+++ CHANGELOG.md
@@ -4,6 +4,212 @@ Updates should follow the [Keep a CHANGELOG](https://keepachangelog.com/) princi
 
 ## [Unreleased][unreleased]
 
+## [0.3.19] - 2022-02-03
+
+### Changed
+
+- Added ClientInterface::getClientTimeout method. Related refactoring so Docksal
+  uses the result of ClientInterface::getDefaultTimeout when running a command.
+
+### Fixed
+
+- Restoring cms:drupal:pre:deploy to run the `composer:install` Deployer task rather
+  than the `composer install` command.
+- Removed functions.php autoload reference.
+
+## [0.3.18] - 2022-02-03
+
+### Fixed
+
+- Temporarily removing automatic deploy failure unlock functionality due to order of operations issue.
+
+## [0.3.17] - 2022-02-03
+
+### Fixed
+
+- Fixing `fill` unsupported function type hint issue.
+
+## [0.3.16] - 2022-02-03
+
+### Changed
+
+- Revising main config to auto-unlock following a failure.
+
+### Fixed
+
+- Revising `cms:drupal:db:backup:create` to allow each DB backup command an hour to execute, by default.
+
+## [0.3.15] - 2022-02-02
+
+### Changed
+
+- New Magento 2 recipe.
+
+## [0.3.14] - 2022-02-01
+
+### Changed
+
+- Revising `cms:drupal:db:update` to allow each DB update command an hour to execute, by default.
+
+## [0.3.13] - 2022-01-27
+
+### Changed
+
+- Revising `releases:cleanup` task so releases may be made writable before removal, if desired.
+
+## [0.3.12] - 2022-01-25
+
+### Changed
+
+- Adding `config.local.php`, `databases.local.php` & `settings.local.php` to list of shared file names.
+
+## [0.3.11] - 2022-01-24
+
+### Changed
+
+- Setting writable_dir_names to an empty array by default.
+- Removing writable_mode & writable_use_sudo default value overrides.
+
+### Fixed
+
+- Resolving release_name setting issue.
+
+## [0.3.10] - 2022-01-24
+
+### Changed
+
+- Set writable_use_sudo to true by default.
+
+### Fixed
+
+- Updated default shared_dir_names value to match writable_dir_names value.
+
+## [0.3.9] - 2022-01-24
+
+### Changed
+
+- Added ability to override config provided by contrib recipes.
+- Set writable_mode to chmod by default.
+
+## [0.3.8] - 2022-01-21
+
+### Fixed
+
+- Added automatic GTID purged variable support detection.
+
+## [0.3.7] - 2022-01-21
+
+### Changed
+
+- Updated VirtualMachine class to load a Virtual Machine Client class that implements an interface.
+
+### Fixed
+
+- Resolved GTID issue preventing databases from being imported.
+
+## [0.3.6] - 2022-01-21
+
+Duplicate of 0.3.5.
+
+## [0.3.5] - 2022-01-21
+
+### Fixed
+
+- Revising Drupal private & temporary file paths to be relative in list of `writable_dir_names`.
+
+## [0.3.4] - 2022-01-20
+
+### Changed
+
+- Defined Drupal private & temporary file paths in default list of `writable_dir_names`.
+
+## [0.3.3] - 2021-12-15
+
+### Changed
+
+- Added the value of `{project}` to the default `{environment}_name` variable value to help prevent
+  host name collisions among similar projects.
+
+## [0.3.2] - 2021-12-15
+
+### Fixed
+
+- Revised Drupal multi-site config to prevent project-specific variables from being overridden.
+- Updated WordPress config to be compatible with newer config default values logic.
+
+## [0.3.1] - 2021-12-10
+
+### Fixed
+
+- Revised relevant recipes to be compatible with 7.0.0-rc.3.
+- Defining CMS `deploy` task _after_ the majority of the recipes are loaded.
+- `fill` function `trim` compatibility bug fixed.
+- Revised `release_name` closure to return a string based on timestamp & branch if necessary.
+
+## [0.3.0] - 2021-12-09
+This release adds support for multiple production webservers. In order for such
+functionality to work, the order of operations for the main `deploy.yaml` file needed
+to be modified. Instead of importing the CMS-related recipe _early_ in the file, it
+_must_ be imported _late_. This allows contrib recipes to _react_ to any config
+that was defined earlier by `deploy.yaml`. The earlier approach only allowed global
+config vars to be overridden.
+
+### Changed
+
+- Updated package to use Deployer 7.0.0 release candidate 3.
+- Added support for multiple production webservers.
+- Revised package to load `config.php` _late_, instead of `config.yml` _early_.
+
+## [0.2.1] - 2021-12-03
+
+### Changed
+
+- Added `skip_db_ops` option to allow skipping of all DB-related operations.
+- Added `skip_db_backup` option to allow skipping of DB backup operations.
+
+#### Drupal
+
+- Added `skip_cache_rebuild` option to allow skipping of cache rebuild operations.
+- Added `skip_config_import` option to allow skipping of config import operations.
+- Added `skip_db_update` option to allow skipping of DB update operations.
+- Added `skip_themes_build` option to allow skipping of theme build operations.
+- Multi-site support for `cms:drupal:cache:rebuild`.
+- Multi-site support for `cms:drupal:config:import`.
+- Multi-site support for `cms:drupal:db:update`.
+
+## [0.2.0] - 2021-11-30
+
+### Changed
+
+#### Drupal
+
+- `shared_file_names` now defaults to a single value referencing an `.env` file. The
+  variable can be overridden if standard-fare Drupal configuration via settings files
+  is preferred. See [unleashedtech/dotenv-drupal](https://github.com/unleashedtech/dotenv-drupal) for more.
+- `shared_dir_names`, `shared_file_names` & `writable_dir_names` now require arrays of
+  paths relative to the release root. Each path will be parsed for placeholders before use.
+- The `cms:drupal:db:backup:create` task now supports creating backups of multi-site apps.
+
+### Fixed
+
+#### Drupal
+
+- Database backup(s) are created at the outset of the `deploy` task to help ensure reliability
+  of the database backup task.
+
+## [0.1.9] - 2021-11-03
+
+### Fixed
+
+- Restoring squashed essential Drupal multi-site variables.
+
+## [0.1.8] - 2021-11-03
+
+### Fixed
+
+- Missing Drupal-init-related import issue fixed.
+- YAML script syntax issues resolved.
+
 ## [0.1.7] - 2021-11-03
 
 ### Changed
@@ -77,7 +283,31 @@ Updates should follow the [Keep a CHANGELOG](https://keepachangelog.com/) princi
 
 **Initial release!**
 
-[unreleased]: https://github.com/unleashedtech/deployer-recipes/compare/0.1.7...main
+[unreleased]: https://github.com/unleashedtech/deployer-recipes/compare/0.3.19...main
+[0.3.19]: https://github.com/unleashedtech/deployer-recipes/compare/0.3.18...0.3.19
+[0.3.18]: https://github.com/unleashedtech/deployer-recipes/compare/0.3.17...0.3.18
+[0.3.17]: https://github.com/unleashedtech/deployer-recipes/compare/0.3.16...0.3.17
+[0.3.16]: https://github.com/unleashedtech/deployer-recipes/compare/0.3.15...0.3.16
+[0.3.15]: https://github.com/unleashedtech/deployer-recipes/compare/0.3.14...0.3.15
+[0.3.14]: https://github.com/unleashedtech/deployer-recipes/compare/0.3.13...0.3.14
+[0.3.13]: https://github.com/unleashedtech/deployer-recipes/compare/0.3.12...0.3.13
+[0.3.12]: https://github.com/unleashedtech/deployer-recipes/compare/0.3.11...0.3.12
+[0.3.11]: https://github.com/unleashedtech/deployer-recipes/compare/0.3.10...0.3.11
+[0.3.10]: https://github.com/unleashedtech/deployer-recipes/compare/0.3.9...0.3.10
+[0.3.9]: https://github.com/unleashedtech/deployer-recipes/compare/0.3.8...0.3.9
+[0.3.8]: https://github.com/unleashedtech/deployer-recipes/compare/0.3.7...0.3.8
+[0.3.7]: https://github.com/unleashedtech/deployer-recipes/compare/0.3.6...0.3.7
+[0.3.6]: https://github.com/unleashedtech/deployer-recipes/compare/0.3.5...0.3.6
+[0.3.5]: https://github.com/unleashedtech/deployer-recipes/compare/0.3.4...0.3.5
+[0.3.4]: https://github.com/unleashedtech/deployer-recipes/compare/0.3.3...0.3.4
+[0.3.3]: https://github.com/unleashedtech/deployer-recipes/compare/0.3.2...0.3.3
+[0.3.2]: https://github.com/unleashedtech/deployer-recipes/compare/0.3.1...0.3.2
+[0.3.1]: https://github.com/unleashedtech/deployer-recipes/compare/0.3.0...0.3.1
+[0.3.0]: https://github.com/unleashedtech/deployer-recipes/compare/0.2.1...0.3.0
+[0.2.1]: https://github.com/unleashedtech/deployer-recipes/compare/0.2.0...0.2.1
+[0.2.0]: https://github.com/unleashedtech/deployer-recipes/compare/0.1.9...0.2.0
+[0.1.9]: https://github.com/unleashedtech/deployer-recipes/compare/0.1.8...0.1.9
+[0.1.8]: https://github.com/unleashedtech/deployer-recipes/compare/0.1.7...0.1.8
 [0.1.7]: https://github.com/unleashedtech/deployer-recipes/compare/0.1.6...0.1.7
 [0.1.6]: https://github.com/unleashedtech/deployer-recipes/compare/0.1.5...0.1.6
 [0.1.5]: https://github.com/unleashedtech/deployer-recipes/compare/0.1.4...0.1.5